+++ examples/RangingAnchor/RangingAnchor.ino
@@ -21,10 +21,11 @@
  * in a custom way, as no MAC features are implemented yet.
  *
  * Complements the "RangingTag" example sketch.
- */
-
-/*
- * TODO weighted average of ranging results based on signal quality
+ *
+ * @todo
+ *  - weighted average of ranging results based on signal quality
+ *  - use enum instead of define
+ *  - move strings to flash (less RAM consumption)
  */
 
 #include <SPI.h>
@@ -36,6 +37,7 @@ const uint8_t PIN_IRQ = 2; // irq pin
 const uint8_t PIN_SS = SS; // spi select pin
 
 // messages used in the ranging protocol
+// TODO replace by enum
 #define POLL 0
 #define POLL_ACK 1
 #define RANGE 2
@@ -61,13 +63,13 @@ DW1000Time timeComputedRange;
 #define LEN_DATA 16
 byte data[LEN_DATA];
 // watchdog and reset period
-unsigned long lastActivity;
-unsigned long resetPeriod = 250;
+uint32_t lastActivity;
+uint32_t resetPeriod = 250;
 // reply times (same on both sides for symm. ranging)
-unsigned int replyDelayTimeUS = 3000;
+uint16_t replyDelayTimeUS = 3000;
 // ranging counter (per second)
-unsigned int successRangingCount = 0;
-unsigned long rangingCountPeriod = 0;
+uint16_t successRangingCount = 0;
+uint32_t rangingCountPeriod = 0;
 float samplingRate = 0;
 
 void setup() {
@@ -203,7 +205,7 @@ void computeRangeSymmetric() {
  */
 
 void loop() {
-    long curMillis = millis();
+    int32_t curMillis = millis();
     if (!sentAck && !receivedAck) {
         // check if inactive
         if (curMillis - lastActivity > resetPeriod) {
